Dispatch actions in corresponding components (instead of parent App) because sometimes the state will change from component to component (e.g. fetch collections of this or that user, or fetch all collections)
