+++ client/src/routes/UsersPage.js
@@ -1,11 +1,20 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import { Link, Outlet, useLocation } from 'react-router-dom';
+import { getUsers } from '../redux/users/usersSlice';
 import Users from '../components/users/Users';
 
 const UsersPage = () => {
+  const dispatch = useDispatch();
   const { pathname } = useLocation();
+
+  useEffect(() => {
+    dispatch(getUsers());
+  }, [dispatch]);
+
   return (
     <>
-      <h3>USERS COMPONENT</h3>
+      <h2>USERS PAGE</h2>
       <Users />
       {pathname === '/users/form'
         ? <Link to="/users">Close</Link>
